+++ torrentfile/TorrentFile.h
@@ -9,19 +9,20 @@
 #include <string>
 #include <vector>
 #include <array>
+#include <cstdint>
 
 struct TorrentFile
 {
     std::string announce;
-    std::array<int, 20> infoHash;
-    std::vector<std::array<int, 20>> pieceHashes;
+    std::array<uint8_t, 20> infoHash;
+    std::vector<std::array<uint8_t, 20>> pieceHashes;
     int pieceLength;
     int length;
     std::string name;
 
     TorrentFile(const std::string& announce,
-                const std::array<int, 20>& infoHash,
-                const std::vector<std::array<int, 20>>& pieceHashes,
+                const std::array<uint8_t, 20>& infoHash,
+                const std::vector<std::array<uint8_t, 20>>& pieceHashes,
                 int pieceLength,
                 int length,
                 const std::string& name);
